pyproject.toml
[project]
name = "breakwater"
version = "0.1.0"
description = "Korean Banking Software Analysis Platform"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "typer>=0.9.0",
    "pyyaml>=6.0",
    "requests>=2.28.0",
    "pycryptodome>=3.23.0",
    "sqlalchemy>=2.0.0",
    "lxml>=4.9.0",
    "cryptography>=45.0.6",
    "pyghidra>=2.2.0",
]

[project.scripts]
breakwater = "breakwater.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/breakwater"]

[tool.pyright]
pythonVersion = "3.13"
venvPath = "."
venv = ".venv"
typeCheckingMode = "standard"
reportUnknownParameterType = true
strictListInference = true
strictDictionaryInference = true
exclude = [".venv", "**/__pycache__"]
include = ["src/breakwater/**/*.py"]
ignore = [".venv/**"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors (PEP8 formatting)
    "F",   # pyflakes (logical errors, unused imports/vars)
    "I",   # isort (import sorting)
    "B",   # flake8-bugbear (common bugs and bad practices)
    "UP",  # pyupgrade (modernize syntax for newer Python versions)
    "SIM", # flake8-simplify (suggest simpler code constructs)
    "TC",  # flake8-type-checking (optimize type checking for runtime)
    "RUF", # Ruff-specific rules (extra linting goodies)
    "N",   # pep8-naming (enforce naming conventions)
]
